requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0

# Celery and Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Email
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# WhatsApp
requests==2.31.0

# AI/ML
openai==1.6.1

# Payment
paystackapi==1.2.6

# API Documentation
drf-yasg==1.21.7

# Environment variables
python-decouple==3.8

# Utils
python-dateutil==2.8.2
pytz==2023.3.post1

# WSGI server
gunicorn==21.2.0

# Development
black==23.12.1
flake8==6.1.0
